% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{palette2color}
\alias{palette2color}
\title{Choose a color palette to get colors}
\usage{
palette2color(palette, n = NULL, modeColor = 1)
}
\arguments{
\item{palette}{A character string specifying the name of the color palette.}
}
\value{
A vector of color hex codes corresponding to the chosen palette.
}
\description{
This function takes a color palette name as input and returns a corresponding list of colors.
}
\examples{
palette2color(palette = "rainbow1")
palette2color(palette = c("red","blue")) # This will return to the original state
}
