% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scale_color_sci}
\alias{scale_color_sci}
\title{Automated Color Setting}
\usage{
scale_color_sci(
  ...,
  palette,
  type = c("discrete", "continuous"),
  modeColor = "auto",
  na.value = "grey50",
  aesthetics = "colour",
  direction = -1
)
}
\arguments{
\item{...}{Additional arguments passed to the scale functions.}

\item{palette}{A vector of colors for the palette.}

\item{type}{Specifies the type of scale: either "discrete" or "continuous".}

\item{modeColor}{Mode of color selection; default is "auto". If "1", colors are applied one by one.}

\item{na.value}{The color used for missing (NA) values; default is "grey50".}

\item{aesthetics}{The aesthetic to map the color scale to, usually "colour".}

\item{direction}{The direction of the color scale; -1 means no change, 1 means reversed.}
}
\value{
A scale function for the color aesthetic.
}
\description{
Select a color palette to apply to ggplot2 plots. If the number of colors in the palette is smaller than required, it will automatically be extended.
}
\examples{

# Discrete data application
 ggplot(mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
   geom_point(size = 4) +
   scale_color_sci(
     palette = "color1",
     type = "discrete",
     modeColor = "auto",
     name = "Cylinders",
     direction = 1
   )

 # Continuous data application
 ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
   geom_tile() +
   scale_fill_sci(
     palette = "color1",
     type = "continuous"
   ) +
   labs(title = "Custom Continuous Color Scale Application")



}
