% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_style1}
\alias{theme_style1}
\title{Custom Scientific Theme}
\usage{
theme_style1(base_size = 14)
}
\arguments{
\item{base_size}{The base font size for the theme. Default is 14.}
}
\value{
A `ggplot2` theme object with custom settings for scientific plots.
}
\description{
A custom theme designed for scientific plots, featuring a clean and professional look suitable for research presentations and publications. It includes white backgrounds, minimal grid lines, and a bold title and axis labels.
}
\examples{
ggplot(mtcars, aes(x = wt, y = mpg,fill=carb)) +
  geom_point() +
  theme_style1(base_size = 14) +
  labs(title = "Car Weight vs MPG")

}
